requirements.txt
Flask
PyMySQL
python-dotenv
Flask-Bcrypt
Flask-Login
requests
